requirements.txt
discord==1.7.3
Pillow==9.1.0
py-cord==2.2.2
requests==2.25.1
requests2==2.16.0